fix(ios): align wrappers with real VeepooBleSDK ObjC signatures
The +Feature.swift bridge split was never compiled against the vendor
framework; these call sites used assumed signatures. Match the headers:
- WatchFace: screenStyle is (int) -> pass Int32(screenIndex)
- Weather: VPWeatherServerHourlyModel.code is int -> use .int32Value
- WomenHealth/WristFlipWake/Sedentary: ...WithXxxModel: imports as (with:)
- Alarms: veepooSDKSettingDevice{New,Text}Alarm(with:) base name + label

Verified: VeepooSDK pod target compiles clean (BUILD SUCCEEDED).

+++ ios/VeepooSDK/VeepooSDKModule+WatchFace.swift
@@ -99,7 +99,7 @@ extension VeepooSDKModule {
     }
     let dial = dialKeyToVpType(extractDialKey(settings))
     peripheralManage.veepooSDKSettingDeviceScreenStyle(
-      screenIndex,
+      Int32(screenIndex),
       settingMode: 1,
       dialType: dial,
       result: { _, _, ok in
